Client et Serveur peuvent communiquer. Gestion des erreurs. C'était marrant

+++ server/src/server/Calculator.java
@@ -1,8 +1,18 @@
 package server;
 
 public class Calculator {
-    public static int operation (int op1, int op2, char op) throws InterruptedException {
-        int result;
+
+    /**
+     * Do the operation asked by the client
+     *
+     * @param op1 rhs operand
+     * @param op2 lhs operand
+     * @param op operator
+     * @return the result of the calculation asked by the client
+     * @throws InterruptedException if the calculation is not supported, throws an exception
+     */
+    public static double operation (double op1, double op2, char op) throws InterruptedException {
+        double result;
 
         switch (op) {
             case '+' :
